sline.c: sline(): Reduces the amount of characters to size - 1
We need to change the meaning of what the `size' parameter of sline()
means, making it like fgets()'s, i.e. "one less than `size'." If we
allow sline() to write chars up to `size', we're allowing for strings
that aren't null-terminated, as '\0' would fall out of bounds.

This way we are guaranteeing that the string is always null-terminated.