Ignoring tabs in input
Tabs are currently ignored by sline in input to avoid the screen and
buffer to get out of sync due to variable tab widths. There doesn't
seem to be any benefit in trying to handle tabs, as sline doesn't
support multi-line editing, where tab-indentation might be of help.
By definining a specific VT_TAB symbol for the internal VT-100 driver,
everything is set in place for implementing some type of user-definable
tab-completion in a future version of sline.